// apps/crm/app/games/games.d.ts
export type currentFeatures = {
  id: string
  createdAt: Date
  updatedAt: Date
  type: string
}

export type currentProviders = {
  id: string
  createdAt: Date
  updatedAt: Date
  type: string
}

export type currentReels = {
  id: string
  createdAt: Date
  updatedAt: Date
  type: string
}

export type currentThemes = {
  id: string
  createdAt: Date
  updatedAt: Date
  type: string
}

export type currentVolatility = {
  id: string
  createdAt: Date
  updatedAt: Date
  type: string
}